refactor: Convert monitoring modules to use Decimal
- Update PreValidator to use Decimal for amount validation
- Convert TransactionParser to use Decimal for swap amounts
- Update WalletPerformance to use Decimal for metrics
- Convert SignalAggregator to use Decimal
- Update RpcPolling to use Decimal where applicable
- Ensures monitoring calculations maintain precision

+++ operator/src/monitoring/rpc_polling.rs
@@ -11,6 +11,7 @@ use crate::monitoring::rate_limiter::RateLimiter;
 use crate::monitoring::rate_limiter::RequestPriority;
 use crate::db::DbPool;
 use anyhow::{Context, Result};
+use rust_decimal::Decimal;
 use solana_client::rpc_client::RpcClient;
 
 /// Transaction information from polling
@@ -20,7 +21,7 @@ pub struct WalletTransaction {
     pub signature: String,
     pub token_address: Option<String>,
     pub direction: Option<String>, // BUY or SELL
-    pub amount_sol: Option<f64>,
+    pub amount_sol: Option<Decimal>,
     pub timestamp: i64,
 }
